feat: sets finances figures to whole numbers of CFP

+++ platform/src/abstractions/Platform.Domain/DataObjects/FinancialPlanDataObject.cs
@@ -16,11 +16,11 @@ public record FinancialPlanDataObject
     [JsonProperty("version")] public int Version { get; set; }
     [JsonProperty("isComplete")] public bool IsComplete { get; set; }
     [JsonProperty("targetContactRatio")] public decimal TargetContactRatio { get; set; }
-    [JsonProperty("totalIncome")] public decimal? TotalIncome { get; set; }
-    [JsonProperty("totalExpenditure")] public decimal? TotalExpenditure { get; set; }
-    [JsonProperty("totalTeacherCosts")] public decimal? TotalTeacherCosts { get; set; }
+    [JsonProperty("totalIncome")] public int? TotalIncome { get; set; }
+    [JsonProperty("totalExpenditure")] public int? TotalExpenditure { get; set; }
+    [JsonProperty("totalTeacherCosts")] public int? TotalTeacherCosts { get; set; }
     [JsonProperty("totalNumberOfTeachersFte")] public decimal? TotalNumberOfTeachersFte { get; set; }
-    [JsonProperty("educationSupportStaffCosts")] public decimal? EducationSupportStaffCosts { get; set; }
+    [JsonProperty("educationSupportStaffCosts")] public int? EducationSupportStaffCosts { get; set; }
     [JsonProperty("useFigures")] public bool? UseFigures { get; set; }
     [JsonProperty("timetablePeriods")] public int? TimetablePeriods { get; set; }
     [JsonProperty("hasMixedAgeClasses")] public bool? HasMixedAgeClasses { get; set; }
